feat: try to refact r1cs
